refactor: move AutoScrollSwitch from General file, don't use cache for npm
