+++ .github/workflows/frontend-review.yml
@@ -25,7 +25,7 @@ jobs:
         uses: actions/setup-node@v3
         with:
           node-version: 20
-          cache: 'npm'
+          # cache: 'npm'
 
       - name: Install dependencies
         run: npm ci
